+++ gdx-vfx/core/src/com/crashinvaders/vfx/VfxManager.java
@@ -21,7 +21,6 @@ import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Pixmap.Format;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.Texture.TextureWrap;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Disposable;
 import com.crashinvaders.vfx.effects.ChainVfxEffect;
@@ -42,14 +41,11 @@ public final class VfxManager implements Disposable {
     /** Maintains a per-frame updated list of enabled effects */
     private final Array<ChainVfxEffect> effectsEnabled = new Array<>(); //TODO Get rid of this.
     private final VfxRenderContext context;
+
     private final VfxPingPongWrapper pingPongWrapper;
-    private final ScreenCaptureHelper captureHelper;
 
+    private boolean capturing = false;
     private boolean disabled = false; //TODO Remove the property.
-//    private boolean capturing = false;  //TODO Remove the property.
-
-    private boolean hasInputScene = false;
-    private boolean hasProcessedScene = false;
 
     private boolean applyingEffects = false;
 
@@ -69,8 +65,6 @@ public final class VfxManager implements Disposable {
 
         // VfxFrameBufferPool will manage both ping-pong VfxFrameBuffer instances for us.
         this.pingPongWrapper = new VfxPingPongWrapper(context.getBufferPool());
-
-        captureHelper = new ScreenCaptureHelper();
     }
 
     @Override
@@ -85,6 +79,10 @@ public final class VfxManager implements Disposable {
 
     public int getHeight() {
         return height;
+    }
+
+    public boolean isCapturing() {
+        return capturing;
     }
 
     public boolean isDisabled() {
@@ -116,25 +114,16 @@ public final class VfxManager implements Disposable {
         return context.getPixelFormat();
     }
 
-    public void setBufferTextureParams(TextureWrap u, TextureWrap v, Texture.TextureFilter min, Texture.TextureFilter mag) {
-        Texture srcTexture = pingPongWrapper.getSrcTexture();
-        srcTexture.setWrap(u, v);
-        srcTexture.setFilter(min, mag);
-        Texture dstTexture = pingPongWrapper.getSrcTexture();
-        dstTexture.setWrap(u, v);
-        dstTexture.setFilter(min, mag);
+    public void setEffectTextureParams(
+            Texture.TextureWrap textureWrapU,
+            Texture.TextureWrap textureWrapV,
+            Texture.TextureFilter textureFilterMin,
+            Texture.TextureFilter textureFilterMag) {
+        this.context.getBufferPool().setTextureParams(textureWrapU, textureWrapV, textureFilterMin, textureFilterMag);
     }
-
-//    public boolean isCapturing() {
-//        return capturing;
-//    }
 
     public boolean isApplyingEffects() {
         return applyingEffects;
-    }
-
-    public boolean hasProcessedScene() {
-        return hasProcessedScene;
     }
 
     /** @return the last active destination frame buffer. */
@@ -149,10 +138,6 @@ public final class VfxManager implements Disposable {
 
     public VfxRenderContext getRenderContext() {
         return context;
-    }
-
-    public ScreenCaptureHelper getCaptureHelper() {
-        return captureHelper;
     }
 
     /**
@@ -195,12 +180,9 @@ public final class VfxManager implements Disposable {
     /** Cleans up the {@link VfxPingPongWrapper}'s buffers with the color specified. */
     public void cleanUpBuffers(Color color) {
         if (applyingEffects) throw new IllegalStateException("Cannot clean up buffers when applying effects.");
-        if (captureHelper.isCapturing()) throw new IllegalStateException("Cannot clean up buffers when capturing a scene.");
+        if (capturing) throw new IllegalStateException("Cannot clean up buffers when capturing a scene.");
 
         pingPongWrapper.cleanUpBuffers(color);
-
-        hasProcessedScene = false;
-        hasInputScene = false;
     }
 
     public void resize(int width, int height) {
@@ -228,84 +210,50 @@ public final class VfxManager implements Disposable {
         }
     }
 
-//    /**
-//     * Starts capturing the scene.
-//     *
-//     * @return true or false, whether or not capturing has been initiated.
-//     * Capturing will fail if the manager is disabled or capturing is already started.
-//     */
-//    public boolean beginCapture() {
-//        if (applyingEffects) {
-//            throw new IllegalStateException("You cannot capture when you're applying the effects.");
-//        }
-//
-//        if (disabled) return false;
-//        if (capturing) return false;
-//
-//        capturing = true;
-//        pingPongWrapper.begin();
-//        return true;
-//    }
-//
-//    /**
-//     * Stops capturing the scene.
-//     * @return false if there was no capturing before that call.
-//     */
-//    public boolean endCapture() {
-//        if (!capturing) throw new IllegalStateException("The capturing is not started. Forgot to call beginCapture()?");
-//
-//        hasCaptured = true;
-//        capturing = false;
-//        pingPongWrapper.end();
-//        return true;
-//    }
+    /** Starts capturing the input buffer. */
+    public void beginInputCapture() {
+        if (applyingEffects) {
+            throw new IllegalStateException("Capture is not available when VfxManager is applying the effects.");
+        }
+        if (capturing) return;
 
-    /** @see VfxManager#useAsInputScene(Texture)  */
-    public void useAsInputScene(VfxFrameBuffer frameBuffer) {
-        useAsInputScene(frameBuffer.getTexture());
+        capturing = true;
+        pingPongWrapper.begin();
+    }
+
+    /** Stops capturing the input buffer. */
+    public void endInputCapture() {
+        if (!capturing) throw new IllegalStateException("The capturing is not started. Forgot to call #beginInputCapture()?");
+
+        capturing = false;
+        pingPongWrapper.end();
+    }
+
+    /** @see VfxManager#useAsInput(Texture)  */
+    public void useAsInput(VfxFrameBuffer frameBuffer) {
+        useAsInput(frameBuffer.getTexture());
     }
 
     /** Sets up a (captured?) source scene that will be used later as an input for effect processing.
      * Updates the effect chain src buffer with the data provided. */
-    public void useAsInputScene(Texture texture) {
-        if (captureHelper.isCapturing()) {
+    public void useAsInput(Texture texture) {
+        if (capturing) {
             throw new IllegalStateException("Cannot set captured input when capture helper is currently capturing.");
+        }
+        if (applyingEffects) {
+            throw new IllegalStateException("Cannot update the input buffer when applying effects.");
         }
 
         context.getBufferRenderer().renderToFbo(texture, pingPongWrapper.getDstBuffer());
-
-        hasInputScene = true;
-        hasProcessedScene = false;
     }
 
-    public void useAsInputScene(ScreenCaptureHelper captureHelper) {
-        if (captureHelper != this.captureHelper) {
-            throw new IllegalStateException("Can only use capture helper obtained though VfxManager#getCaptureHelper() method of this instance.");
-        }
-        if (captureHelper.isCapturing()) {
-            throw new IllegalStateException("Capture helper is in capturing stage. Forgot to call CaptureHelper#endCapture() ?");
-        }
-        if (!captureHelper.isResultCaptured()) {
-            throw new IllegalStateException("Capture helper doesn't have a captured result. " +
-                    "You should call CaptureHelper#startCapture() and CaptureHelper#endCapture() " +
-                    "before you can provide the capture helper as an input for the effect processing.");
-        }
-
-        // Capture helper uses VfxManager#pingPongWrapper#bufDst frame buffer to capture the scene.
-        // So we don't need to do anything else to use it as an input.
-        hasInputScene = true;
-        hasProcessedScene = false;
-        captureHelper.resetState();
-    }
-
-    /** Applies the effect chain, if there is one. */
+    /** Applies the effect chain. */
     public void applyEffects() {
-        if (captureHelper.isCapturing()) {
+        if (capturing) {
             throw new IllegalStateException("You should call VfxManager.endCapture() before applying the effects.");
         }
 
         if (disabled) return;
-        if (!hasInputScene) return;
 
         Array<ChainVfxEffect> effectChain = updateEnabledEffectList();
 
@@ -320,9 +268,10 @@ public final class VfxManager implements Disposable {
             Gdx.gl.glDisable(GL20.GL_CULL_FACE);
             Gdx.gl.glDisable(GL20.GL_DEPTH_TEST);
 
-            // Render the effect chain.
-            pingPongWrapper.swap(); // Swap buffers to get the input scene in the src buffer.
+            pingPongWrapper.swap(); // Swap buffers to get the input buffer in the src buffer.
             pingPongWrapper.begin();
+
+            // Render the effect chain.
             for (int i = 0; i < count; i++) {
                 ChainVfxEffect effect = effectChain.get(i);
                 effect.render(context, pingPongWrapper);
@@ -338,18 +287,14 @@ public final class VfxManager implements Disposable {
             if (blendingEnabled) {
                 Gdx.gl.glDisable(GL20.GL_BLEND);
             }
-
-            hasProcessedScene = true;
         }
         applyingEffects = false;
     }
 
     public void renderToScreen() {
-        if (captureHelper.isCapturing()) {
+        if (capturing) {
             throw new IllegalStateException("You should call endCapture() before rendering the result.");
         }
-        if (disabled) return;
-        if (!hasInputScene) return;
 
         // Enable blending to preserve buffer's alpha values.
         if (blendingEnabled) { Gdx.gl.glEnable(GL20.GL_BLEND); }
@@ -358,11 +303,9 @@ public final class VfxManager implements Disposable {
     }
 
     public void renderToScreen(int x, int y, int width, int height) {
-        if (captureHelper.isCapturing()) {
+        if (capturing) {
             throw new IllegalStateException("You should call endCapture() before rendering the result.");
         }
-        if (disabled) return;
-        if (!hasInputScene) return;
 
         // Enable blending to preserve buffer's alpha values.
         if (blendingEnabled) { Gdx.gl.glEnable(GL20.GL_BLEND); }
@@ -371,11 +314,9 @@ public final class VfxManager implements Disposable {
     }
 
     public void renderToFbo(VfxFrameBuffer output) {
-        if (captureHelper.isCapturing()) {
-            throw new IllegalStateException("You should call VfxManager.endCapture() before rendering the result.");
+        if (capturing) {
+            throw new IllegalStateException("You should call endCapture() before rendering the result.");
         }
-        if (disabled) return;
-        if (!hasInputScene) return;
 
         // Enable blending to preserve buffer's alpha values.
         if (blendingEnabled) { Gdx.gl.glEnable(GL20.GL_BLEND); }
@@ -394,58 +335,5 @@ public final class VfxManager implements Disposable {
             }
         }
         return effectsEnabled;
-    }
-
-    public class ScreenCaptureHelper {
-
-        private boolean capturing = false;
-        private boolean resultCaptured = false;
-
-        public boolean isCapturing() {
-            return capturing;
-        }
-
-        public boolean isResultCaptured() {
-            return resultCaptured;
-        }
-
-        /**
-         * Starts capturing the scene.
-         */
-        public void beginCapture() {
-            if (capturing) return;
-
-            if (applyingEffects) {
-                throw new IllegalStateException("You cannot capture when you're applying the effects.");
-            }
-
-            capturing = true;
-            pingPongWrapper.begin();
-        }
-
-        /**
-         * Stops capturing the scene.
-         */
-        public void endCapture() {
-            if (!capturing) throw new IllegalStateException("The capturing is not started. Forgot to call beginCapture()?");
-
-            hasInputScene = true;
-            capturing = false;
-            pingPongWrapper.end();
-
-            resultCaptured = true;
-        }
-
-        public VfxFrameBuffer getCapturedResult() {
-            return pingPongWrapper.getDstBuffer();
-        }
-
-        protected void resetState() {
-            if (capturing) {
-                throw new IllegalStateException("Cannot reset the state when in capturing state.");
-            }
-            capturing = false;
-            resultCaptured = false;
-        }
     }
 }